move routing into its own folder
