+++ src/app/app.component.html
@@ -1,3 +1,3 @@
 <app-header></app-header>
 
-<h1>Placeholder1</h1>
+<router-outlet></router-outlet>
