+++ src/app/routing/interceptors/token-interceptor/token.interceptor.ts
