+++ src/app/routing/resolvers/fetch-all-quizzes/fetch-all-quizzes.resolver.ts
@@ -3,7 +3,7 @@ import { ResolveFn } from '@angular/router';
 import { QuizService } from '../../../features/quiz-service/quiz.service';
 import { IQuizList, order, sort } from '../../../../types/others/lists.types';
 import { Observable } from 'rxjs';
-import { paramsBuilder } from '../../params-builder/params-builder';
+import { paramsBuilder } from '../../../util/params-builder/params-builder';
 
 /**
  * Retrieves a catalogue of all quizzes so that they can be loaded before navigating
@@ -12,18 +12,16 @@ import { paramsBuilder } from '../../params-builder/params-builder';
  * @param state 
  * @returns 
  */
-export const fetchUserQuizzesResolver: ResolveFn<IQuizList> = (route, state): Observable<IQuizList> => {
+export const fetchAllQuizzesResolver: ResolveFn<IQuizList> = (route, state): Observable<IQuizList> => {
   const quizService = inject(QuizService);
 
-  const userId = route.paramMap.get('id') || '';
   const page = Number(route.queryParamMap.get('page'));
   const sort = route.queryParamMap.get('sort') as sort;
   const order = route.queryParamMap.get('order') as order;
 
   const params = paramsBuilder(page, sort, order);
 
-  return quizService.getUserQuizzes(
-    userId,
+  return quizService.getAllQuizzes(
     Number(params.get('page')),
     params.get('sort') as sort,
     params.get('order') as order,
